moved search button to right side
this was accomplished by making a seperate definition to right justify
the search tab, as well as adding a parent div over the both of them. I
then had to set the fill of the equally spaced tabs to 70% to keep the
search button from getting pushed onto the next line. Should only break
when a display has fewer than 340 pixels wide (and even then, only
"about" is going to dissapear)